refactor (NormalFrame) : step2 - 프레임의 결과값 RollingResult로 객체 분리
