+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,50 +1,18 @@
 package bowling.domain.frame;
 
-import bowling.domain.exception.InvalidThrowBallException;
-
 public class NormalFrame {
-    private static final int KNOCKED_DOWN_PIN_COUNT_DEFAULT = -1;
-    private static final int KNOCKED_DOWN_PIN_COUNT_STRIKE = 10;
-
-    private Integer firstRollingResult;
-    private Integer secondRollingResult;
+    RollingResult rollingResult;
 
     public NormalFrame() {
-        this.firstRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
-        this.secondRollingResult = KNOCKED_DOWN_PIN_COUNT_DEFAULT;
+        rollingResult = RollingResult.init();
     }
 
     public void rollingBall(int pinCount) {
-        if (isFrameEnd()) {
-            throw new InvalidThrowBallException();
-        }
-
-        if ((firstRollingResult == KNOCKED_DOWN_PIN_COUNT_DEFAULT)) {
-            this.firstRollingResult = pinCount;
-            return;
-        }
-
-        this.secondRollingResult = pinCount;
+        rollingResult.roll(pinCount);
     }
 
     public boolean isRollable() {
-        if (isFrameEnd()) {
-            return false;
-        }
-
-        if (firstRollingResult == KNOCKED_DOWN_PIN_COUNT_STRIKE) {
-            return false;
-        }
-
-        return true;
+        return rollingResult.isRollable();
     }
 
-    private boolean isFrameEnd() {
-        if (firstRollingResult != KNOCKED_DOWN_PIN_COUNT_DEFAULT &&
-                secondRollingResult != KNOCKED_DOWN_PIN_COUNT_DEFAULT) {
-            return true;
-        }
-
-        return false;
-    }
 }
